module7-task3: fix tests
