+++ project/src/store/index.ts
@@ -1,7 +1,7 @@
 import { configureStore } from '@reduxjs/toolkit';
 
 import { createAPI } from '../api/api';
-import { rootReducer } from './rootReducer';
+import { rootReducer } from './root.reducer';
 
 const api = createAPI();
 export const store = configureStore({
